% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicants.R
\docType{data}
\name{applicants}
\alias{applicants}
\title{Applicants}
\format{
\subsection{\code{applicants}}{

A data frame with 3 rows and 8 columns:
\describe{
\item{Student}{ A letter to indentify the person}
\item{MatI}{Mark in Mathematics I}
\item{MatII}{Mark in Mathematics II}
\item{MatIII}{Mark in Mathematics III}
\item{StatI}{Mark in Statistics I}
\item{StatII}{Mark in Statistics II}
\item{EcoI}{Mark in Economics I}
\item{EcoII}{Mark in Economics II}
}
}
}
\source{
Paper
}
\usage{
applicants
}
\description{
A subset of data from the paper of LLamazares
}
\keyword{datasets}
