% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicants_B.R
\docType{data}
\name{applicants_B}
\alias{applicants_B}
\title{applicants_B}
\format{
\subsection{\code{who}}{

A data frame with 5 rows and 6 columns:
\describe{
\item{Candidate}{An identificator for each candidate}
\item{P1}{Assessments assigned by the professor P1}
\item{P1}{Assessments assigned by the professor P2}
\item{P1}{Assessments assigned by the professor P3}
\item{P1}{Assessments assigned by the academic A1}
\item{P1}{Assessments assigned by the academic A2}
\item{P1}{Assessments assigned by the head of the departmen H}
}
}
}
\source{
\url{https://onlinelibrary.wiley.com/doi/full/10.1002/int.21913}
}
\usage{
applicants_B
}
\description{
A data frame containing the assessments carried out by a panel formed by 6 evaluators.
}
\details{
An academic department wishes to hire faculty that can contribute significantly
to research, teaching, industry engagement and service and leadership. Additionaly,
the applicant's research should match at least one of the field of research conducted in the department.
The panel members supply their overall scores  which need to be aggregated into
an overall score for ranking of the applicants.

This case was published as un illustrative example by Beliakov, G.(2018)
"Comparing Apples and Oranges: The Weighted OWA Function" in International Journal
of Intelligent Systems, vol. 33, issue 5.
}
\keyword{datasets}
