% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capacity_OWA.R
\name{capacity_OWA}
\alias{capacity_OWA}
\title{The capacity (or fuzzy measure) associated with an Ordered Weighted Average (OWA)}
\usage{
capacity_OWA(weights_w)
}
\arguments{
\item{weights_w}{A numerical vector of positive numbers. It gives the weights to use in an Ordered Weighted Average (OWA)}
}
\value{
A numeric vector giving the capacity corresponding to each subset from the set formed from the numbers 1,2,..., n
}
\description{
The capacity (or fuzzy measure) associated with an Ordered Weighted Average (OWA)
}
\examples{
capacity_OWA(c(0.2, 0.3, 0.3, 0.2))
}
