% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capacity_SUOWA.R
\name{capacity_SUOWA}
\alias{capacity_SUOWA}
\title{The capacity (or fuzzy measure) associated with a SUOWA operator}
\usage{
capacity_SUOWA(weights_p, weights_w, uninorm = "uninorm_maxmin")
}
\arguments{
\item{weights_p}{A numerical vector of positive numbers. It gives the weights to use in the weighted mean (WEM)}

\item{weights_w}{A numerical vector of positive numbers. It gives the weights to use in the Ordered Weighted Average (OWA)}

\item{uninorm}{The name of the uninom nedeed to computue the capacity. At the moment, only uninom_maxmin is avaliable}
}
\value{
A numeric matrix giving the capacity corresponding to each subset of the set formed from the numbers 1,2,..., n, for the SUOWA operator and the corresponding  to the WEM and the OWA that the SUOWA operator generalized
}
\description{
The capacity (or fuzzy measure) associated with a SUOWA operator
}
\examples{
capacity_SUOWA(weights_p = 1:3, weights_w = 1:3)
}
