% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capacity_WEM.R
\name{capacity_WEM}
\alias{capacity_WEM}
\title{The capacity (or fuzzy measure) associated with a Weighted Mean (WEM)}
\usage{
capacity_WEM(weights_p)
}
\arguments{
\item{weights_p}{A numerical vector of positive numbers. It gives the weights to use in a weighted mean (WEM)}
}
\value{
A numeric vector giving the capacity corresponding to each subset of the set formed from the numbers 1,2,..., n
}
\description{
The capacity (or fuzzy measure) associated with a Weighted Mean (WEM)
}
\examples{
capacity_WEM(1:3)
capacity_WEM(c(0.25, 0.5, 0.25))
}
