% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capacity_winsorized_weighted_mean.R
\name{capacity_winsorized_weighted_mean}
\alias{capacity_winsorized_weighted_mean}
\title{The capacity (or fuzzy measure) associated with the (r,s)-fold Winsorized weighted mean}
\usage{
capacity_winsorized_weighted_mean(
  n,
  r,
  s = r,
  alpha = NULL,
  beta = alpha,
  weights = NULL
)
}
\arguments{
\item{n}{A positive integer. It represents the dimension of the vector to apply winsorized mean}

\item{r}{A positive integer. The number of observations to be winsorized from the lower end of x before the mean is computed}

\item{s}{A positive integer. The number of observations to be winsorized from the upper end of x before the mean is computed}

\item{alpha}{A real number between 0 and 0.5. It gives the proportion of observations to be winsorized from the lower end of x before the mean is computed}

\item{beta}{A real number between 0 and 0.5. It gives the proportion of observations to be winsorized from the upper end of x before the mean is computed}

\item{weights}{A numerical vector of weights with the same length as x. It gives the weights to use in the Winsorized weighted mean}
}
\value{
An object of class capacity.
}
\description{
The capacity (or fuzzy measure) associated with the (r,s)-fold Winsorized weighted mean
}
\examples{
capacity_winsorized_weighted_mean(n=6, r=2)
}
