% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/favor_winsorized_weighted_mean.R
\name{favor_winsorized_weighted_mean}
\alias{favor_winsorized_weighted_mean}
\title{A favor indice to measure the behaviour of of the winsorized weighted means}
\usage{
favor_winsorized_weighted_mean(weights, r, s = r, alpha = NULL, beta = alpha)
}
\arguments{
\item{weights}{A numerical vector of weights}

\item{r}{An integer. The number of observations to be winsorized from the lower end of x before the mean is computed}

\item{s}{An integer. The number of observations to be winsorized from the upper end of x before the mean is computed}

\item{alpha}{The fraction (0 to 0.5) of observations to be winsorized from the lower end of x before the mean is computed}

\item{beta}{The fraction (0 to 0.5) of observations to be winsorized from the upper end of x before the mean is computed}
}
\value{
A numerical vector of the same length of weights. A favor indice measures the degree with which the behaviour of a criterion (coordenate of the vector to aggregate) is like a favor
}
\description{
A favor indice to measure the behaviour of of the winsorized weighted means
}
\examples{
favor_winsorized_weighted_mean(weights = c(rep(0.2,10), 0.5,0.5), r=2, s=2)
}
