% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orness_winsorized_weighted_mean.R
\name{orness_winsorized_weighted_mean}
\alias{orness_winsorized_weighted_mean}
\title{The orness degree of a Winsorized weighted mean}
\usage{
orness_winsorized_weighted_mean(weights, r, s = r, alpha = NULL, beta = alpha)
}
\arguments{
\item{weights}{A numerical vector of weights}

\item{r}{An integer. The number of observations to be winsorized from the lower end of x before the mean is computed}

\item{s}{An integer. The number of observations to be winsorized from the upper end of x before the mean is computed}

\item{alpha}{The fraction (0 to 0.5) of observations to be winsorized from the lower end of x before the mean is computed}

\item{beta}{The fraction (0 to 0.5) of observations to be winsorized from the upper end of x before the mean is computed}
}
\value{
A numercial value. The orness degree of the Winsorized weighted mean required
}
\description{
The orness degree of a Winsorized weighted mean
}
\examples{
orness_winsorized_weighted_mean(weights = rep(0.2,10), r=2, s=2)
}
