% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Shapley_values.R
\name{plot_Shapley_values}
\alias{plot_Shapley_values}
\title{Plot of Shapley values}
\usage{
plot_Shapley_values(w, r = NULL, s = r)
}
\arguments{
\item{w}{A numerical vector of positive values.}

\item{r}{An integer. The number of observations to be winsorized from the lower end of x before the mean is computed}

\item{s}{An integer. The number of observations to be winsorized from the upper end of x before the mean is computed}
}
\value{
Plot of weights and Shapley values
}
\description{
Plot of Shapley values
}
\examples{
plot_Shapley_values(w=c(rep(0.2,3), rep(0.1,4)), r=2)
}
