% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_set.R
\name{power_set}
\alias{power_set}
\title{Title}
\usage{
power_set(n, set = NULL, sort = TRUE)
}
\arguments{
\item{n}{A positive integer. It represents the size of the set}

\item{set}{A vector containing the elements od a set}

\item{sort}{A logical denoting if the subsets are sorted by cardinality}
}
\value{
A list containing all the subsets (the power set) togetherr their cardinality. If sort is TRUE (the dafault), sorted by cardinality.
}
\description{
Title
}
\examples{
power_set(4)
power_set(set=c("a", "b", "c","d"))
}
