% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapley_values_winsorized_weighed_mean.R
\name{Shapley_values_winsorized_weighed_mean}
\alias{Shapley_values_winsorized_weighed_mean}
\title{Shapley values associated with the (r,s)-fold Winsorized weighted mean}
\usage{
Shapley_values_winsorized_weighed_mean(
  n,
  r = 1,
  s = r,
  alpha = NULL,
  beta = alpha,
  w = NULL
)
}
\arguments{
\item{n}{A positive integer. It represents the dimension of the vector to apply winsorized mean}

\item{r}{A positive integer. The number of observations to be winsorized from the lower end of x before the mean is computed}

\item{s}{A positive integer. The number of observations to be winsorized from the upper end of x before the mean is computed}

\item{alpha}{A real number between 0 and 0.5. It gives the proportion of observations to be winsorized from the lower end of x before the mean is computed}

\item{beta}{A real number between 0 and 0.5. It gives the proportion of observations to be winsorized from the upper end of x before the mean is computed}

\item{w}{A numerical vector of weights with the same length as x. It gives the weights to use in the Winsorized weighted mean}
}
\value{
The Shapley values for a winsorized weighted mean for a vecotor of length n computed using the arguments given
}
\description{
Shapley values associated with the (r,s)-fold Winsorized weighted mean
}
\examples{
x<-1:20
winsorized_weighted_mean(x,r=2)
Shapley_values_winsorized_weighed_mean(n=20, r=2)
}
