% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapley_wm.R
\name{shapley_wm}
\alias{shapley_wm}
\title{Shapley values for a winsorized weighted mean}
\usage{
shapley_wm(n, r = 1, s = r, alpha = NULL, beta = alpha, w = NULL)
}
\arguments{
\item{n}{dimension of the vector to apply winsorized mean}

\item{r}{An integer. The number of observations to be winsorized from the lower end of x before the mean is computed}

\item{s}{An integer. The number of observations to be winsorized from the upper end of x before the mean is computed}

\item{alpha}{The fraction (0 to 0.5) of observations to be winsorized from the lower end of x before the mean is computed}

\item{beta}{The fraction (0 to 0.5) of observations to be winsorized from the upper end of x before the mean is computed}

\item{w}{A numerical vector of weights the same length as x giving the weights to use for elements of x}
}
\value{
The Shapley values for a winsorized weighted mean for a vecotor of length n computed using the arguments given
}
\description{
Shapley values for a winsorized weighted mean
}
\examples{
x<-1:20
winsorized_mean(x,r=2)
shapley_wm(n=20, r=2)
}
