% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SUOWA.R
\name{SUOWA}
\alias{SUOWA}
\title{It computes the SUOWA aggegated value of a vector}
\usage{
SUOWA(x, mu_SUOWA, final_weights = FALSE)
}
\arguments{
\item{x}{A numerical vector to be aggegate}

\item{mu_SUOWA}{A numerical vector giving the capacity (fuzzy measure) to be used}

\item{final_weights}{A logical value to require the specific final weights used during the computation of the SUOWA aggregated value}
}
\value{
A numeric value with the aggegated value obtained using the SUOWA operator. When it is require, this function returns the ordered initial vector toghether with the corresponding weights used for the computation
}
\description{
It computes the SUOWA aggegated value of a vector
}
\examples{
we<-c(0, 0.5, 0.5, 0 )
pe<- c(0.5, 0.2, 0.2, 0.1)
mu <- capacity_SUOWA(weights_p = pe, weights_w = we, uninorm = "uninom_max_min")[, "mu_SUOWA"]
x <- c(10, 5, 4, 6)
SUOWA(x, mu_SUOWA = mu)
SUOWA(x, mu_SUOWA = mu, final_weights =TRUE)
}
