% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uninorm_maxmin.R
\name{uninorm_maxmin}
\alias{uninorm_maxmin}
\title{The max - min uninorm}
\usage{
uninorm_maxmin(x, y, n)
}
\arguments{
\item{x}{A number between 0 and 1}

\item{y}{A number between 0 and 1}

\item{n}{A natural number. 1/n is the neutral element of this max-min uninorm}
}
\value{
A numeric vector giving the capacity corresponding to each subset from the set formed from the numbers 1,2,..., n
}
\description{
The max - min uninorm
}
\examples{
uninorm_maxmin(0.02,0.04,10)
}
