% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veto_winsorized_weighted_mean.R
\name{veto_winsorized_weighted_mean}
\alias{veto_winsorized_weighted_mean}
\title{A veto indice to measure the behaviour of of the winsorized weighted means}
\usage{
veto_winsorized_weighted_mean(weights, r, s = r, alpha = NULL, beta = alpha)
}
\arguments{
\item{weights}{A numerical vector of weights}

\item{r}{An integer. The number of observations to be winsorized from the lower end of x before the mean is computed}

\item{s}{An integer. The number of observations to be winsorized from the upper end of x before the mean is computed}

\item{alpha}{The fraction (0 to 0.5) of observations to be winsorized from the lower end of x before the mean is computed}

\item{beta}{The fraction (0 to 0.5) of observations to be winsorized from the upper end of x before the mean is computed}
}
\value{
A numerical vector of the same length of weights. A veto indice measures the degree with which the behaviour of a criterion (coordenate of the vector to aggregate) is like a veto
}
\description{
A veto indice to measure the behaviour of of the winsorized weighted means
}
\examples{
veto_winsorized_weighted_mean(weights = rep(0.2,10), r=2, s=1)
}
