% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights_from_Shapley_values.R
\name{weights_from_Shapley_values}
\alias{weights_from_Shapley_values}
\title{Weigths to use in a Winsorized weighted mean calculated from Shapley vales given}
\usage{
weights_from_Shapley_values(Shapley_values, r, s, alpha = NULL, beta = alpha)
}
\arguments{
\item{Shapley_values}{A numerical vector with the Shapley values given}

\item{r}{An integer. The number of observations to be winsorized from the lower end of x before the mean is computed}

\item{s}{An integer. The number of observations to be winsorized from the upper end of x before the mean is computed}

\item{alpha}{The fraction (0 to 0.5) of observations to be winsorized from the lower end of x before the mean is computed}

\item{beta}{The fraction (0 to 0.5) of observations to be winsorized from the upper end of x before the mean is computed}
}
\value{
A vector of weights
}
\description{
Weigths to use in a Winsorized weighted mean calculated from Shapley vales given
}
\examples{
v<- c(rep(0.2,3),rep(0.1, 4))
weights_from_Shapley_values(Shapley_values=v , r=1, s=0)

}
