% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winsorized_weighted_mean.R
\name{winsorized_weighted_mean}
\alias{winsorized_weighted_mean}
\title{Winsorized weighted mean}
\usage{
winsorized_weighted_mean(
  x,
  r = 1,
  s = r,
  alpha = NULL,
  beta = alpha,
  weights = NULL,
  na.rm = FALSE,
  ...
)
}
\arguments{
\item{x}{A numerical vector}

\item{r}{An integer. The number of observations to be winsorized from the lower end of x before the mean is computed}

\item{s}{An integer. The number of observations to be winsorized from the upper end of x before the mean is computed}

\item{alpha}{The fraction (0 to 0.5) of observations to be winsorized from the lower end of x before the mean is computed}

\item{beta}{The fraction (0 to 0.5) of observations to be winsorized from the upper end of x before the mean is computed}

\item{weights}{A numerical vector of weights the same length as x giving the weights to use for elements of x}

\item{na.rm}{a logical evaluating to TRUE or FALSE indicating whether NA values should be stripped before the computation proceeds}

\item{...}{arguments to be passed to or from methods}
}
\value{
The winsorized weighted mean computed using the arguments
}
\description{
Winsorized weighted mean
}
\examples{
winsorized_weighted_mean(x=1:20,r=2)
}
